Split utils.js, move chart tooltip to UtilizationCard
  - utils.js split up to smaller units focused on specific kinds of
    tasks (cpu-topology, validation, vm-icons)

  - moved the UtilizationCard tooltip function to a helper file
    in `.../cards/UtilizationCard`
