+++ src/components/utils/index.js
@@ -1,1 +1,4 @@
+export * from './cpu-topology'
 export * from './utils'
+export * from './validation'
+export * from './vm-icons'
